Issue #14 (Scrollbars on main page) : resolved
Adjusted the CSS on the page to resolve scrollbar on main page issue.

Actions taken:
-Adjusted main page CSS into globals.css

-Put the footer into a layout to be visable on all pages

-Added home link on footer to be able to return to main page

Additional note:
This is my first pull request on a repo that was not my own. It would be awesome to have even the smallest bit of my code be out there. You are a great help in the learning process.